+++ include/core/app.h
@@ -3,10 +3,11 @@
 
 #include "core/errorHandler.h"
 #include "core/appContext.h"
+#include "expressionEngine/functionManager.h"
 #include "textRenderer/text.h"
 
 enum reh_error_code_e ra_appInit(struct ra_app_context_t *ctx);
-enum reh_error_code_e ra_appRenderFrame(struct ra_app_context_t *ctx, struct rtr_character_t *chars);
+enum reh_error_code_e ra_appRenderFrame(struct ra_app_context_t *ctx, struct rtr_character_t *chars, struct ree_function_manager_t *functions);
 void ra_appShutdown(struct ra_app_context_t *ctx, const char *msg);
 
 #endif // APP_H