+++ include/expressionEngine/parser/functionParser.h
@@ -5,6 +5,7 @@
 #include "expressionEngine/functionManager.h"
 
 enum reh_error_code_e ree_testLexerParser(void);
-enum reh_error_code_e ree_parseFunctionDefinition(char* definition, struct ree_function_t *function);
+enum reh_error_code_e ree_implicitMultiplication(struct ree_token_t **tokens, int *tokenCount, int *tokenCapacity);
+enum reh_error_code_e ree_parseFunction(char* definition, struct ree_function_t *function);
 
 #endif // FUNCTION_PARSER_H